完善getAccessTokenInfo方法

+++ Java/spring-boot/14.spring boot auth2.0/spring-boot-oauth-server/src/main/java/com/glitter/spring/boot/bean/AccessTokenInfo.java
@@ -12,7 +12,7 @@ public class AccessTokenInfo extends BaseBean {
     private String token_type;
 
     /** 过期时间,单位为秒 */
-    private Integer expires_in;
+    private Long expires_in;
 
     /** 更新令牌 */
     private String refresh_token;
